Custom Tool Development/providers/dns_enum.py
import dns.resolver

def enum(domain):
    print(f"  [*] Enumerating DNS records for {domain} (Placeholder)...")
    dns_data = {}
    record_types = ['A', 'MX', 'TXT', 'NS']
    for rtype in record_types:
        try:
            answers = dns.resolver.resolve(domain, rtype)
            dns_data[rtype] = [str(r) for r in answers]
        except dns.resolver.NoAnswer:
            dns_data[rtype] = "No records found"
        except dns.resolver.NXDOMAIN:
            dns_data[rtype] = "Domain does not exist"
            break
        except Exception as e:
            dns_data[rtype] = f"Error: {e}"
    return dns_data
