Custom Tool Development/requirements.txt
requests
PyYAML
python-whois
dnspython
scapy
python-nmap
